comments

+++ Exercise01/Models/Person.cs
@@ -1,3 +1,4 @@
+// Define the namespace where the Person class is located.
 namespace Howest.MCT.Ex1.Models;
 public class Person
 {
@@ -8,7 +9,7 @@ public class Person
         LastName = lastname;
     }
 
-    public Person(string name, string lastname, byte age, string address, string email, string phone)
+    public Person(string name, string lastname, int age, string address, string email, string phone)
     {
         Name = name;
         LastName = lastname;
@@ -19,6 +20,7 @@ public class Person
     }
 
     // FUNCTIONS
+    // Override the ToString() method 
     public override string ToString()
     {
         return $"Person Information: \n Name: {Name} \n Last Name: {LastName} \n Age: {Age} \n Address: {Address} \n Email: {Email} \n Phone: {Phone}";
@@ -27,9 +29,8 @@ public class Person
     // PROPERTIES
     public string Name { get; set; }
     public string LastName { get; set; }
-    public byte Age { get; set; }
+    public int Age { get; set; }
     public string Address { get; set; }
     public string Email { get; set; }
-    public string Phone { get; set; }
-    
+    public string Phone { get; set; }   
 }
